refactor: extract TaskList component from TaskListScreen
- Create TaskList as a presentational component for list rendering
- TaskListScreen now focuses on state management and business logic
- Improves separation of concerns and component reusability
- Enables TaskList to be used in other contexts (e.g., search results)
